// utils/api/generated/models/get-ticket-history200-response-data-tickets-inner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Chatbot
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface GetTicketHistory200ResponseDataTicketsInner
 */
export interface GetTicketHistory200ResponseDataTicketsInner {
  /**
   *
   * @type {string}
   * @memberof GetTicketHistory200ResponseDataTicketsInner
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof GetTicketHistory200ResponseDataTicketsInner
   */
  subject: string;
  /**
   *
   * @type {string}
   * @memberof GetTicketHistory200ResponseDataTicketsInner
   */
  description: string;
  /**
   *
   * @type {string}
   * @memberof GetTicketHistory200ResponseDataTicketsInner
   */
  answer?: string | null;
  /**
   *
   * @type {boolean}
   * @memberof GetTicketHistory200ResponseDataTicketsInner
   */
  isRated: boolean;
  /**
   *
   * @type {string}
   * @memberof GetTicketHistory200ResponseDataTicketsInner
   */
  departmentId: string;
  /**
   *
   * @type {string}
   * @memberof GetTicketHistory200ResponseDataTicketsInner
   */
  createdAt: string;
  /**
   *
   * @type {string}
   * @memberof GetTicketHistory200ResponseDataTicketsInner
   */
  updatedAt: string;
  /**
   *
   * @type {string}
   * @memberof GetTicketHistory200ResponseDataTicketsInner
   */
  status: string;
  /**
   *
   * @type {string}
   * @memberof GetTicketHistory200ResponseDataTicketsInner
   */
  code: string;
}